seperation of layouts into multiple xml files addition of CodeEditorLayout as a managing tool for child views (line numbers, divider, edittext)
