+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <de.markusressel.kodeeditor.library.view.CodeEditorView
+    <de.markusressel.kodeeditor.library.view.CodeEditorLayout
         android:id="@+id/codeEditorView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
