+++ library/src/main/java/de/markusressel/kodeeditor/library/view/CodeEditorView.kt
@@ -2,35 +2,22 @@ package de.markusressel.kodeeditor.library.view
 
 import android.content.Context
 import android.content.res.TypedArray
-import android.graphics.Rect
 import android.graphics.drawable.Drawable
-import android.os.Build
 import android.support.annotation.AttrRes
 import android.support.annotation.ColorInt
 import android.support.annotation.StringRes
 import android.support.annotation.StyleableRes
 import android.support.v4.content.ContextCompat
 import android.support.v4.view.ViewCompat
-import android.text.Layout
 import android.util.AttributeSet
-import android.util.Log
 import android.util.TypedValue
 import android.view.LayoutInflater
-import android.view.MotionEvent
 import android.view.View
 import android.widget.LinearLayout
 import android.widget.TextView
-import com.jakewharton.rxbinding2.widget.RxTextView
 import com.otaliastudios.zoom.ZoomLayout
-import com.trello.rxlifecycle2.kotlin.bindToLifecycle
 import de.markusressel.kodeeditor.library.R
 import de.markusressel.kodeeditor.library.syntaxhighlighter.SyntaxHighlighter
-import io.reactivex.Observable
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.rxkotlin.subscribeBy
-import io.reactivex.schedulers.Schedulers
-import java.util.concurrent.TimeUnit
-import kotlin.math.roundToInt
 
 /**
  * Code Editor that allows pinch-to-zoom, line numbers etc.
@@ -43,16 +30,6 @@ open class CodeEditorView : ZoomLayout {
     internal lateinit var contentLayout: LinearLayout
 
     /**
-     * The view displaying line numbers
-     */
-    internal lateinit var lineNumberView: TextView
-
-    /**
-     * The divider between line numbers and text editor
-     */
-    internal lateinit var dividerView: View
-
-    /**
      * The actual EditText
      */
     internal lateinit var editTextView: CodeEditText
@@ -61,9 +38,6 @@ open class CodeEditorView : ZoomLayout {
      * Set to true to force the width of the CodeEditorView to it's parents width
      */
     private var forceParentWidth = false
-
-    private var moveWithCursorEnabled = false
-    private var currentLineCount = -1
 
     constructor(context: Context) : super(context) {
         initialize(null, 0)
@@ -108,29 +82,6 @@ open class CodeEditorView : ZoomLayout {
         val a = context
                 .obtainStyledAttributes(attrs, R.styleable.CodeEditorView, defStyleAttr, 0)
 
-        val lineNumberTextColor = a
-                .getColor(R.styleable.CodeEditorView_cev_lineNumbers_textColor, R.attr.cev_lineNumbers_textColor, android.R.attr.textColorPrimary)
-        lineNumberView
-                .setTextColor(lineNumberTextColor)
-
-        val lineNumberBackgroundColor = a
-                .getColor(R.styleable.CodeEditorView_cev_lineNumbers_backgroundColor, R.attr.cev_lineNumbers_backgroundColor, android.R.attr.windowBackground)
-        lineNumberView
-                .setBackgroundColor(lineNumberBackgroundColor)
-
-
-        val dividerEnabled = a
-                .getBoolean(R.styleable.CodeEditorView_cev_divider, true)
-        dividerView
-                .visibility = when (dividerEnabled) {
-            true -> View.VISIBLE
-            else -> View.GONE
-        }
-        val dividerColor = a
-                .getColor(R.styleable.CodeEditorView_cev_divider_color, R.attr.cev_divider_color, android.R.attr.textColorPrimary)
-        dividerView
-                .setBackgroundColor(dividerColor)
-
         val editTextBackgroundColor = a
                 .getColor(R.styleable.CodeEditorView_cev_editor_backgroundColor, R.attr.cev_editor_backgroundColor, android.R.attr.windowBackground)
         editTextView
@@ -142,13 +93,6 @@ open class CodeEditorView : ZoomLayout {
 
     private fun inflateViews(inflater: LayoutInflater) {
         contentLayout = inflater.inflate(R.layout.view_code_editor__inner_layout, null) as LinearLayout
-
-        lineNumberView = contentLayout.findViewById(R.id.codeLinesView) as TextView
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            lineNumberView.hyphenationFrequency = Layout.HYPHENATION_FREQUENCY_NONE
-        }
-
-        dividerView = contentLayout.findViewById(R.id.divider) as View
 
         editTextView = contentLayout.findViewById(R.id.codeEditText) as CodeEditText
         editTextView
@@ -213,61 +157,8 @@ open class CodeEditorView : ZoomLayout {
                     contentLayout
                             .layoutParams = params
                 }
-                updateLineNumbers(editTextView.lineCount)
             }
         }
-
-        setOnTouchListener { view, motionEvent ->
-            when (motionEvent.action) {
-                MotionEvent.ACTION_MOVE -> {
-                    moveWithCursorEnabled = false
-                }
-            }
-            false
-        }
-
-        editTextView
-                .setOnClickListener {
-                    moveWithCursorEnabled = true
-                }
-
-        Observable
-                .interval(250, TimeUnit.MILLISECONDS)
-                .filter { moveWithCursorEnabled }
-                .bindToLifecycle(this)
-                .subscribeBy(onNext = {
-                    try {
-                        moveScreenWithCursorIfNecessary()
-                    } catch (e: Throwable) {
-                        Log
-                                .e(TAG, "Error moving screen with cursor", e)
-                    }
-                }, onError = {
-                    Log
-                            .e(TAG, "Unrecoverable error while moving screen with cursor", it)
-                })
-
-        RxTextView
-                .textChanges(editTextView)
-                .debounce(50, TimeUnit.MILLISECONDS)
-                .filter {
-                    moveWithCursorEnabled = true
-                    editTextView.lineCount != currentLineCount
-                }
-                .subscribeOn(Schedulers.computation())
-                .observeOn(AndroidSchedulers.mainThread())
-                .bindToLifecycle(this)
-                .subscribeBy(onNext = {
-                    try {
-                        updateLineNumbers(editTextView.lineCount)
-                    } catch (e: Throwable) {
-                        Log
-                                .e(TAG, "Error updating line numbers", e)
-                    }
-                }, onError = {
-                    Log
-                            .e(TAG, "Unrecoverable error while updating line numbers", it)
-                })
     }
 
     private fun setMinimumDimensions() {
@@ -282,67 +173,6 @@ open class CodeEditorView : ZoomLayout {
                 .minimumHeight = parentHeight
         contentLayout
                 .minimumWidth = parentWidth
-    }
-
-    private fun moveScreenWithCursorIfNecessary() {
-        val pos = editTextView
-                .selectionStart
-        val layout = editTextView
-                .layout
-
-        if (layout != null) {
-            val line = layout
-                    .getLineForOffset(pos)
-            val baseline = layout
-                    .getLineBaseline(line)
-            val ascent = layout
-                    .getLineAscent(line)
-            val x = layout
-                    .getPrimaryHorizontal(pos)
-            val y = (baseline + ascent)
-                    .toFloat()
-
-            val zoomLayoutRect = Rect()
-            getLocalVisibleRect(zoomLayoutRect)
-
-            val transformedX = x * realZoom + panX * realZoom + lineNumberView.width * realZoom
-            val transformedY = y * realZoom + panY * realZoom
-
-            if (!zoomLayoutRect.contains(transformedX.roundToInt(), transformedY.roundToInt())) {
-
-                var newX = panX
-                var newY = panY
-
-                if (transformedX < zoomLayoutRect.left || transformedX > zoomLayoutRect.right) {
-                    newX = -x
-                }
-
-                if (transformedY < zoomLayoutRect.top || transformedY > zoomLayoutRect.bottom) {
-                    newY = -y
-                }
-
-                moveTo(zoom, newX, newY, false)
-            }
-        }
-    }
-
-    private fun updateLineNumbers(lines: Int) {
-        currentLineCount = lines
-
-        val linesToDraw = if (lines < MIN_LINES) {
-            MIN_LINES
-        } else {
-            lines
-        }
-
-        val sb = StringBuilder()
-        for (i in 1..linesToDraw) {
-            sb
-                    .append("$i:\n")
-        }
-        lineNumberView
-                .text = sb
-                .toString()
     }
 
     /**
@@ -361,7 +191,6 @@ open class CodeEditorView : ZoomLayout {
                 .setText(text, TextView.BufferType.EDITABLE)
         editTextView
                 .refreshSyntaxHighlighting()
-        updateLineNumbers(editTextView.lineCount)
     }
 
     /**
@@ -377,8 +206,7 @@ open class CodeEditorView : ZoomLayout {
      */
     @Suppress("unused")
     fun setSyntaxHighlighter(syntaxHighlighter: SyntaxHighlighter) {
-        editTextView
-                .syntaxHighlighter = syntaxHighlighter
+        editTextView.syntaxHighlighter = syntaxHighlighter
     }
 
     companion object {
