+++ library/src/main/res/layout/view_code_editor__divider.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<View xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/divider"
+    android:layout_width="1dp"
+    android:layout_height="match_parent" />
