+++ library/src/main/res/layout/view_code_editor__editor.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<de.markusressel.kodeeditor.library.view.CodeEditorView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" />
