+++ library/src/main/res/layout/view_code_editor__linenumbers.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/codeLinesView"
+    android:layout_width="wrap_content"
+    android:layout_height="match_parent"
+    android:gravity="right|top"
+    android:paddingLeft="4dp"
+    android:paddingTop="10dp"
+    android:textSize="12sp"
+    android:typeface="monospace" />
